requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-environ==0.10.0
django-redis==5.3.0
PyMySQL==1.1.0
celery==5.2.6
django-celery-results==2.3.1
django-celery-beat==2.2.1
redis==4.3.1
pandas==2.2.2
djangorestframework==3.13.1
cryptography
scikit-learn==1.5.1
insightface==0.7.3
opencv-python==4.10.0.84
onnxruntime==1.19.0
django-cors-headers==4.4.0